requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.0.3.1
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-Migrate==1.1.0
Flask-Moment==0.2.1
Flask-SQLAlchemy==2.1
Flask-Script==0.6.6
Flask-WTF==0.9.4
Jinja2==2.7.1
Mako==0.9.1
MarkupSafe==0.18
SQLAlchemy==1.1.5
WTForms==1.0.5
Werkzeug==0.10.4
alembic==0.6.2
blinker==1.3
itsdangerous==0.23
pymysql==0.6.6
flask-httpauth
scrapy==1.0.3
dicttoxml==1.6.6
pyyaml==3.11
validate-email==1.3
twisted
xmltodict==0.9.2



